refactor(contributions): use Decimal in contribution schemas

+++ src/contributions/schemas.py
@@ -1,12 +1,13 @@
 from pydantic import BaseModel, Field, EmailStr, ConfigDict
 import uuid
+from decimal import Decimal
 from src.contributions.models import PaymentStatus
 from datetime import datetime
 
 
 class ContributionCreate(BaseModel):
     campaign_id: uuid.UUID
-    amount: float = Field(gt=0, description="Contribution amount in GHS")
+    amount: Decimal = Field(gt=0, description="Contribution amount in GHS")
     contributor_email: EmailStr
     contributor_name: str = Field(min_length=2, max_length=100)
 
@@ -17,7 +18,7 @@ class ContributionResponse(BaseModel):
     user_id: uuid.UUID | None
     contributor_email: str
     contributor_name: str
-    amount: float
+    amount: Decimal
     payment_status: PaymentStatus
     paystack_reference: str | None
     created_at: datetime
